+++ src/main/java/com/imsweb/datagenerator/provider/facility/rule/NameRule.java
@@ -22,7 +22,6 @@ public class NameRule extends ProviderDataGeneratorRule {
         super(ID, "Facility Name");
     }
 
-
     @Override
     public void execute(Map<String, String> provider, ProviderDataGeneratorOptions options) {
         if (options == null || options.getState() == null)
@@ -39,9 +38,9 @@ public class NameRule extends ProviderDataGeneratorRule {
         provider.put("addressPostalCode", dto.getAddressPostalCode());
         provider.put("addressTelephone", dto.getAddressTelephone());
         for (int i = 0; i < 15; i++) {
-            if (dto.getSpecialty(i) != null)
-                if (dto.getSpecialty(i).trim().equals(""))
-                    provider.put("specialty" + (i + 1), dto.getSpecialty(i));
+            String thisSpecialty = dto.getSpecialty(i);
+            if ((thisSpecialty != null) && (!thisSpecialty.trim().equals("")))
+                provider.put("specialty" + (i + 1), thisSpecialty);
         }
     }