+++ src/test/java/com/imsweb/datagenerator/provider/physician/rule/NameRuleTest.java
@@ -3,7 +3,14 @@
  */
 package com.imsweb.datagenerator.provider.physician.rule;
 
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+import org.junit.Assert;
 import org.junit.Test;
+
+import com.imsweb.datagenerator.provider.ProviderDataGeneratorOptions;
 
 public class NameRuleTest {
 
@@ -11,31 +18,19 @@ public class NameRuleTest {
 
     @Test
     public void testExecute() {
-        /*
-        for (String sex : Arrays.asList("1", "2")) {
-            for (String race1 : Arrays.asList("01", "02", "03", "04", "05", "06", "07", "32", "96", "97", "98", "99")) {
-                for (String hispanicOrigin : Arrays.asList("0", "1", "7", "8", "9")) {
-                    Map<String, String> rec = new HashMap<>();
-                    rec.put("sex", sex);
-                    rec.put("race1", race1);
-                    rec.put("spanishHispanicOrigin", hispanicOrigin);
-                    _rule.execute(rec, new ArrayList<Map<String, String>>(), null);
 
-                    // verify pattern of names
-                    Assert.assertTrue(rec.get("nameFirst").matches("^[a-zA-Z]+$"));
-                    Assert.assertTrue(rec.get("nameMiddle").matches("^[a-zA-Z]+$") || rec.get("nameMiddle").matches("^[A-Z]\\.$"));
-                    Assert.assertTrue(rec.get("nameLast").matches("^[a-zA-Z]+$"));
-                    Assert.assertTrue(rec.get("namePrefix").matches("^[\\w \\.]*$"));
-                    Assert.assertTrue(rec.get("nameSuffix").matches("^[\\w \\.]*$"));
+        for (String state : Arrays.asList("WA", "ND", "MD")) {
 
-                    // verify male patient was not given a maiden name, else verify female maiden name if present
-                    if (sex.equals("1"))
-                        Assert.assertTrue(rec.get("nameMaiden").isEmpty());
-                    else
-                        Assert.assertTrue(rec.get("nameMaiden").isEmpty() || rec.get("nameMaiden").matches("^[a-zA-Z]+$"));
-                }
-            }
+            Map<String, String> provider = new HashMap<>();
+            ProviderDataGeneratorOptions options = new ProviderDataGeneratorOptions();
+            options.setState(state);
+            _rule.execute(provider, options);
+
+            Assert.assertTrue(provider.get("npi").matches("\\d{10}"));
+            Assert.assertTrue(provider.get("nameLast").length() > 0);
+            Assert.assertTrue(provider.get("nameFirst").length() > 0);
+            Assert.assertTrue(provider.get("addressState").equals(state));
+
         }
-        */
     }
 }